% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.R
\name{as.choices}
\alias{as.choices}
\title{Choices and alternatives}
\usage{
as.choices(rankings, names = FALSE)
}
\arguments{
\item{rankings}{a matrix of dense rankings.}

\item{names}{logical: if \code{TRUE} use the object names in the returned
\code{"choices"} object, else use object indices.}
}
\description{
coerce a matrix of rankings to a list of choices, alternatives, and
rankings. The choices and the corresponding alternatives is the
exchangeable bit of the Plackett-Luce with ties.
}
\examples{
R <- matrix(c(1, 2, 0, 0,
              4, 1, 2, 3,
              2, 1, 1, 1,
              1, 2, 3, 0,
              2, 1, 1, 0,
              1, 0, 3, 2), nrow = 6, byrow = TRUE)
colnames(R) <- c("apple", "banana", "orange", "pear")
actual_choices <- as.choices(R, names = TRUE)
coded_choices <- as.choices(R, names = FALSE)
attr(coded_choices, "objects")

## Coercion to tibble is straightforwards
tibble::as.tibble(coded_choices)
}
