% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{as.edgelist}
\alias{as.edgelist}
\title{Convert Objects to an Edge List}
\usage{
as.edgelist(x, ...)
}
\arguments{
\item{x}{a \code{\link{rankings}} object.}

\item{...}{further argumjents passed to/from methods.}
}
\value{
a two column matrix
}
\description{
A generic function to convert an object to an edge list of the corresponding
graph. The package provides a method for \code{"rankings"} objects.
}
\details{
For a \code{"rankings"} object an edge is defined for consecutive pairs in
the ranking, with the winner specifying the \sQuote{from} vertex and the loser
specifying the \sQuote{to} vertex. For tied items an edge is defined for
all possible pairs between the winning and the losing group. Pairs are
ordered by rank combination, i.e. all pairs between items ranked first and
items ranked second, for all rankings, then all pairs between items ranked
second and items ranked third, etc.
}
\examples{
X <- matrix(c(2, 1, 2, 1, 2,
              3, 2, 0, 0, 1,
              1, 0, 2, 2, 3), nrow = 3, byrow = TRUE)
as.edgelist(as.rankings(X))

}
