% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{Decode orderings}
\usage{
decode(orderings, items)
}
\arguments{
\item{orderings}{A data frame of coded orderings or an object that can be
coerced to such a data frame.}

\item{items}{A data frame of the items in each ranking, or a vector of
common items. The elements names are used as the key to the code if the
coded values in \code{orderings} are character, otherwise the indices of
the elements are used.}
}
\value{
A data frame with the coded values replaced by the item labels.#'
}
\description{
Decode orderings
}
\examples{
# orderings of up to 3 items coded a A, B, C
orderings <- data.frame(Rank1 = c("A", "B"),
                        Rank2 = c("C", "A"),
                        Rank3 = c("B", NA),
                        stringsAsFactors = FALSE)
items <- data.frame(A = c("banana", "apple"),
                    B = c("orange", "pear"),
                    C = c("apple", NA),
                    stringsAsFactors = FALSE)
decode(orderings, items)

# orderings with ties of up to 3 items, coded 1:3
orderings <- data.frame(Rank1 = c(1, 3),
                        Rank2 = I(list(c(2, 3), 2)),
                        Rank3 = c(NA, 1),
                        stringsAsFactors = FALSE)
items <- data.frame(A = c("banana", "apple"),
                    B = c("orange", "pear"),
                    C = c("apple", "orange"),
                    stringsAsFactors = FALSE)
decode(orderings, items)

# same items in each comparison
items <- c(A = "banana", B = "orange", C = "pear")
decode(orderings, items)
}
