% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_rankings.R
\name{grouped_rankings}
\alias{grouped_rankings}
\alias{as.grouped_rankings}
\alias{as.grouped_rankings.paircomp}
\alias{format.grouped_rankings}
\title{Grouped Rankings Object}
\usage{
grouped_rankings(rankings, index, ...)

as.grouped_rankings(x, ...)

\method{as.grouped_rankings}{paircomp}(x, ...)

\method{format}{grouped_rankings}(x, max = 2, width = 20, ...)
}
\arguments{
\item{rankings}{a \code{\link{rankings}} object or an object that can be
coerced by \code{as.rankings}.}

\item{index}{a numeric vector of length equal to the number of rankings
specifying the subject for each ranking.}

\item{...}{additional arguments passed on to \code{\link{as.rankings}}
by \code{grouped_rankings} or \code{as.grouped_rankings}; unused by
\code{format}.}

\item{x}{an object that can be coerced to a \code{"grouped_rankings"} object
for \code{as.group_rankings}, or a \code{"grouped_rankings"} object for
\code{format}.}

\item{max}{the maximum number of rankings to format per subject.}

\item{width}{the maximum width in number of characters to format each
ranking.}
}
\value{
an object of class \code{"grouped_rankings"}.
}
\description{
Create an object of class \code{"grouped_rankings"} which associates a
subject index with an object of class \code{"rankings"}. This allows the
rankings to be linked to subject covariates as the basis for
model-based recursive partitioning, see \code{\link{pltree}}.
}
\examples{

# ungrouped rankings
R <- as.rankings(matrix(c(1, 2, 0, 0, 0,
                          0, 2, 1, 0, 0,
                          0, 0, 1, 2, 0,
                          2, 1, 0, 0, 0,
                          0, 1, 2, 3, 0), ncol = 5, byrow = TRUE))
length(R)
R

# grouped rankings
G <- grouped_rankings(R, rep(1:2, c(3, 2)))
length(G)
## by default up to 2 rankings are shown per subject, "..." indicates if
## there are further rankings
G
print(G, max = 1)
}
\seealso{
\code{\link{pltree}}
}
