% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlackettLuce-deprecated.R, R/grouped_rankings.R
\name{PlackettLuce-deprecated}
\alias{PlackettLuce-deprecated}
\alias{grouped_rankings}
\title{Deprecated functions in package \pkg{PlackettLuce}}
\usage{
grouped_rankings(rankings, index, ...)
}
\description{
The functions listed below are deprecated and will be defunct in
the near future. When possible, alternative functions with similar
functionality are also mentioned.
}
\section{grouped_rankings}{

\code{grouped_rankings()} has been renamed \code{\link[=group_rankings]{group_rankings()}}.
}

\keyword{internal}
