% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pladmmfit.R
\name{pladmmfit}
\alias{pladmmfit}
\title{Wrapper around pladmm for Model-based Recursive Partitioning}
\usage{
pladmmfit(
  y,
  x = NULL,
  start = NULL,
  weights = NULL,
  offset = NULL,
  ...,
  estfun = FALSE,
  object = FALSE
)
}
\arguments{
\item{y}{a \code{"\link{grouped_rankings}"} object giving the rankings to
model.}

\item{x}{model matrix specifying the dependence of item worth on covariates.}

\item{start}{starting values for the coefficients.}

\item{weights}{not yet implemented.}

\item{offset}{unused.}

\item{...}{additional arguments passed to \code{pladmm}.}

\item{estfun}{logical. If \code{TRUE} the empirical estimating functions
(score/gradient contributions) are returned.}

\item{object}{logical. If \code{TRUE} the fitted model is returned.}
}
\value{
a list with elements
\item{coefficients}{ model coefficients. }
\item{objfun}{ the negative log-likelihood. }
\item{estfun}{ if \code{estfun} the empirical estimating functions. }
\item{object}{ if \code{object} the fitted model. }
}
\description{
This is a wrapper around \code{pladmm} as required by
\code{\link[partykit]{mob}} for model-based recursive partitioning. It is
not intended for general use.
}
\examples{

if (require(prefmod)){
  data(salad)
  # data.frame of rankings for salad dressings A B C D
  # 1 = most tart, 4 = least tart
  salad[1:3,]

  # create data frame of corresponding features
  # (acetic and gluconic acid concentrations in salad dressings)
  features <- data.frame(salad = LETTERS[1:4],
                         acetic = c(0.5, 0.5, 1, 0),
                         gluconic = c(0, 10, 0, 10))

  # create a grouped rankings object, grouping the rankings into 2 groups
  G <- group(as.rankings(salad), rep(1:2, nrow(salad)/2))

  # pladmmfit() gives the same results as pladmm()
  mod1 <- pladmmfit(G, model.matrix(~ acetic + gluconic, data = features),
                    rho = 8)
  mod1$coefficients
  -mod1$objfun

  mod2 <- pladmm(salad, ~ acetic + gluconic, data = features, rho = 8)
  coef(mod2)
  logLik(mod2)

}
}
\keyword{internal}
