% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvcalc.R
\name{qvcalc}
\alias{qvcalc}
\alias{qvcalc.default}
\alias{qvcalc.PlackettLuce}
\title{Quasi Variances for Model Coefficients}
\usage{
qvcalc(object, ...)

\method{qvcalc}{default}(object, factorname = NULL, coef.indices = NULL,
  labels = NULL, dispersion = NULL, estimates = NULL, modelcall = NULL,
  ...)

\method{qvcalc}{PlackettLuce}(object, ref = NULL, ...)
}
\arguments{
\item{object}{An object acceopted by \code{\link[qvcalc]{qvcalc}} or a
\code{"PlackettLuce"} object.}

\item{...}{Further arguments passed to or from other methods.}

\item{factorname}{See \code{\link[qvcalc]{qvcalc}}.}

\item{coef.indices}{See \code{\link[qvcalc]{qvcalc}}.}

\item{labels}{See \code{\link[qvcalc]{qvcalc}}.}

\item{dispersion}{See \code{\link[qvcalc]{qvcalc}}.}

\item{estimates}{See \code{\link[qvcalc]{qvcalc}}.}

\item{modelcall}{See \code{\link[qvcalc]{qvcalc}}.}

\item{ref}{an integer or character string specifying the reference item (for which log ability will be set to zero). If \code{NULL} the first item is used.}
}
\description{
Computes a set of quasi variances (and corresponding quasi standard errors)
for estimated model coefficients relating to the levels of a categorical
(i.e., factor) explanatory variable. For details of the method see Firth
(2000), Firth (2003) or Firth and de Menezes (2004). Quasi variances
generalize and improve the accuracy of “floating absolute risk”
(Easton et al., 1991).
}
