% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.soc.R
\name{read.soi}
\alias{read.soi}
\title{Title}
\usage{
read.soi(file)
}
\arguments{
\item{file}{}
}
\value{

}
\description{
Title
}
\examples{
# url for preflib data in the "Election Data" category
preflib <- "http://www.preflib.org/data/election/"

# strict orderings of 6 random cities from 36 in total
cities <- read.soi(file.path(preflib, "cities/ED-00034-00000001.soi"))

# strict orderings of drivers in the 1961 F1 races
# 8 races with 17 to 34 drivers in each
f1 <- read.soi(file.path(preflib, "f1/ED-00010-00000001.soi"))
}
