% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_orderings.R
\name{recode_orderings}
\alias{recode_orderings}
\title{Title}
\usage{
recode_orderings(orderings, items,
  code = sort(unique(unlist(orderings))))
}
\arguments{
\item{code}{}
}
\description{
Title
}
\examples{
orderings <- data.frame(Rank1 = c("A", "B"),
                        Rank2 = c("C", "A"),
                        Rank3 = c("B", NA),
                        stringsAsFactors = FALSE)
items <- data.frame(ItemA = c("banana", "apple"),
                    ItemB = c("orange", "pear"),
                    ItemC = c("apple", NA),
                    stringsAsFactors = FALSE)
recode_orderings(orderings, items)

orderings <- data.frame(Rank1 = c("A", "C"),
                        Rank2 = I(list(c("B", "C"), "B")),
                        Rank3 = c(NA, "A"),
                        stringsAsFactors = FALSE)
items <- data.frame(ItemA = c("banana", "apple"),
                    ItemB = c("orange", "pear"),
                    ItemC = c("apple", "orange"),
                    stringsAsFactors = FALSE)
}
