% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.rankings.R
\name{aggregate.rankings}
\alias{aggregate.rankings}
\alias{freq}
\title{Aggregate rankings}
\usage{
\method{aggregate}{rankings}(x, ...)

freq(x)
}
\arguments{
\item{x}{A \code{\link[=rankings]{"rankings"}} or \code{"aggregated_rankings"} object.}

\item{...}{Additional arguments, currently unused.}
}
\description{
Aggregate rankings, returning an \code{"aggregated_rankings"} object of the
unique rankings and their frequencies. The frequencies can be extracted via
the function \code{freq()}.
}
