% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.PlackettLuce}
\alias{coef.PlackettLuce}
\title{Extract Coefficient of Plackett Luce Model}
\usage{
\method{coef}{PlackettLuce}(object, ref = NULL, type = "all", ...)
}
\arguments{
\item{object}{an object of class "PlackettLuce" as returned by \code{PlackettLuce}.}

\item{ref}{an integer or character string specifying the reference item (for which log ability will be set to zero). If \code{NULL} the first item is used.}

\item{type}{the type of coefficients to return: one of \code{"ties"}, \code{"abilities"} or \code{"all"}.}

\item{...}{additional arguments, currently ignored.}
}
\description{
Extract Coefficient of Plackett Luce Model
}
