% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.PlackettLuce}
\alias{fitted.PlackettLuce}
\title{Fitted probabilities for PlackettLuce objects}
\usage{
\method{fitted}{PlackettLuce}(object, aggregate = TRUE, free = TRUE, ...)
}
\arguments{
\item{object}{a \code{"PlackettLuce"} object as returned by
\code{\link{PlackettLuce}}.}

\item{...}{further arguments passed to method (ignored).}
}
\description{
Fitted probabilities for all choice/alternative combinations in the data
}
\seealso{
as.choices
}
