% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pladmm.R
\name{pladmm}
\alias{pladmm}
\title{Fit a Plackett-Luce Model with Linear Predictor for Log-worth}
\usage{
pladmm(rankings, X, rho = 1, n_iter = 500, rtol = 1e-04)
}
\arguments{
\item{rankings}{a \code{"\link{rankings}"} object, or an object that can be
coerced by \code{as.rankings}.}

\item{X}{a model matrix specifying the linear model for each item, as
produced by \code{\link[=model.matrix]{model.matrix()}}.}

\item{rho}{the penalty parameter in the panalized likelihood, see details.}

\item{n_iter}{the maximum number of iterations (also for inner loops).}

\item{rtol}{the convergence tolerance (also for inner loops)}
}
\description{
Fit a Plackett-Luce model where the log-worth is predicted by a linear
function of covariates. The rankings may be partial
(each ranking completely ranks a subset of the items), but ties are not
supported.
}
\note{
This is a prototype function and the user interface is planned to
change in upcoming versions of PlackettLuce.
}
\examples{

if (require(prefmod)){
  data(salad)
  # data.frame of rankings for salad dressings A B C D
  # 1 = most tart, 4 = least tart
  salad[1:3,]

  # create data frame of corresponding features
  # (acetic and gluconic acid concentrations in salad dressings)
  features <- data.frame(salad = LETTERS[1:4],
                         acetic = c(0.5, 0.5, 1, 0),
                         gluconic = c(0, 10, 0, 10))

  # create design matrix with features
  salad_X <- model.matrix(~ acetic + gluconic, data = features)

  # fit Plackett-Luce model based on covariates
  res_PLADMM <- pladmm(salad, salad_X, rho = 8)
  ## coefficients
  coef(res_PLADMM)
  ## worth
  res_PLADMM$pi
  ## worth as predicted by linear function
  drop(exp(salad_X \%*\% coef(res_PLADMM)))

}

}
