% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLtree.R
\name{pltree}
\alias{pltree}
\title{Plackett-Luce Trees}
\usage{
pltree(formula, data, subset, na.action, cluster, ref = NULL, ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted, of the form
\code{y ~ x1 + ... + xn} where \code{y} should be an object of class \code{\link{grouped_rankings}} and \code{x1}, \dots, \code{xn} are used as partitioning variables.}

\item{data}{an optional data frame containing the variables in the model.}

\item{subset}{A specification of the rows to be used, passed to
\code{\link{model.frame}}.}

\item{na.action}{how NAs are treated, passed to \code{\link{model.frame}}.}

\item{cluster}{an optional vector of cluster IDs to be employed for clustered
covariances in the parameter stability tests, see \code{\link{mob}}.}

\item{ref}{an integer or character string specifying the reference item (for which log ability will be set to zero). If NULL the first item is used.}

\item{...}{additional arguments passed to \code{\link{PlackettLuce}}}
}
\value{
An object of class \code{"pltree"} inheriting from \code{"bttree"}
and \code{"modelparty"}.
}
\description{
Recursive partitioning based on Plackett-Luce models.
}
\details{
Plackett-Luce trees are an application of model-based recursive partitioning
(implemented in \code{\link[partykit]{mob}}) to Plackett-Luce models for
rankings. The partitioning is based on attributes of the subjects ranking the
items. Thus the response should be a
\code{\link{grouped_rankings}} object that groups rankings by subject. This
may be included in a data frame alongside the subject covariates.

Various methods are provided for \code{"pltree"} objects, most of them
inherited from \code{"mob"} objects (e.g. \code{print}, \code{summary}), or
\code{"bttree"} objects (\code{itempar}, \code{plot}). \code{itempar}
extracts the abilities or item parameters from the Plackett-Luce models in
each node of the tree using \code{\link{itempar.PlackettLuce}}. The plot
method employs the \code{\link[psychotree]{node_btplot}}
panel-generating function.
}
\examples{
# Bradley-Terry example

if (require(psychotree)){
    ## Germany's Next Topmodel 2007 data
    data("Topmodel2007", package = "psychotree")
    ## convert paircomp object to grouped rankings
    R <- as.grouped_rankings(Topmodel2007$preference)

    ## fit partition model based on all variables except preference
    ## currently gives warnings because PlackettLuce does not accept weights
    ## but all weights are 1 here
    tm_tree <- pltree(R ~ ., data = Topmodel2007[, -1], minsize = 5)

    ## plot shows abilities constrained to sum to 1
    plot(tm_tree, abbreviate = 1, yscale = c(0, 0.5))
    ## instead show log-abilities with Anja as reference (need to used index)
    plot(tm_tree, abbreviate = 1, worth = FALSE, ref = 6,
         yscale = c(-1.5, 2.2))

    ## log-abilities, zero sum contrast
    itempar(tm_tree, log = TRUE)
    ## abilities with Anja as reference
    itempar(tm_tree, ref = "Anja")
}
}
\seealso{
\code{\link[psychotree]{bttree}}
}
