% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.PlackettLuce}
\alias{simulate.PlackettLuce}
\title{Simulate from \code{PlackettLuce} fitted objects}
\usage{
\method{simulate}{PlackettLuce}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{an object representing a fitted model.}

\item{nsim}{number of response vectors to simulate.  Defaults to \code{1}.}

\item{seed}{an object specifying if and how the random number
    generator should be initialized (\sQuote{seeded}).\cr
    For the "lm" method, either \code{NULL} or an integer that will be
    used in a call to \code{set.seed} before simulating the response
    vectors.  If set, the value is saved as the \code{"seed"} attribute
    of the returned value.  The default, \code{NULL} will not change the
    random generator state, and return \code{\link{.Random.seed}} as the
    \code{"seed"} attribute, see \sQuote{Value}.
  }

\item{...}{additional optional arguments.}
}
\value{
An object of class \code{\link{rankings}} of the same dimension as \code{object$rankings}
}
\description{
Simulate from \code{PlackettLuce} fitted objects
}
\examples{
R <- matrix(c(1, 2, 0, 0,
              4, 1, 2, 3,
              2, 1, 1, 1,
              1, 2, 3, 0,
              2, 1, 1, 0,
              1, 0, 3, 2), nrow = 6, byrow = TRUE)
colnames(R) <- c("apple", "banana", "orange", "pear")
mod <- PlackettLuce(R)
simulate(mod, 5)

s1 <- simulate(mod, 3, seed = 112)
s2 <- simulate(mod, 2, seed = 112)

identical(s1[1:2], s2[1:2])

}
